requirements.txt
python-binance==1.0.15
python-dotenv==0.19.1
